refactoring + Enum add

+++ src/test/java/pojo/petPojo/PetPOJO.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
@@ -13,7 +14,7 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 @JsonInclude(NON_NULL)
 public class PetPOJO {
 
-    Integer id;
+    BigInteger id;
     CategoryPOJO category;
     String name;
     List<String> photoUrls;
